De/select annotation via map for show/hideCallout
Problem: This means that if an annotation is displayed by calling
showCallout(), it can't be deselected by tapping away on the map. It
can only be dismissed by explicitly calling hideCallout().

Analysis: When showCallout() is called on an annotation, it isn't
displayed through the map via selectAnnotation:. As a result, the
annotation is never deselected (and the didDeselectAnnotationView:
delegate method is never called), because the map didn't know it was
selected in the first place.

Fix:
* In showCallout()/hideCallout(), use
  selectAnnotation:/deselectAnnotation: through the map.
* In showCalloutView:/hideCalloutView:, don't call setSelected: on the
  MKAnnotationView. Quoting from the link below: "You should not call
  this method directly. An MKMapView object calls this method in
  response to user interactions with the annotation."

References:
* https://developer.apple.com/library/prerelease/ios/documentation/MapKit/Reference/MKAnnotationView_Class/index.html#//apple_ref/occ/instm/MKAnnotationView/setSelected:animated: