Pass through props even when map isn't ready
We encountered an issue at a hackathon where props (and therefore styles) was not being passed through: this means that if the map isn't ready (`this.state.isReady` is false), it is invisible unless it was given a width and height by its parent. This is an issue because on genymotion, where maps aren't available, isReady never becomes true and we can never see the error message.