Targets the current React Native.
I hear this is the wrong way to do this.  I'm not sure what the right
way is.  All I know is, without this change, the project doesn't
compile on Android and RN 0.29.2.

Feels wrong to specify the React Native version number in here, when at
the NPM level, we target the "latest" React Native dependency with a *.