Enable control of Google Maps Marker tracksViewChanges property.
Fixes issue #1031, enabling use of custom marker images and views in
Google Maps on iOS and android.

As detailed in the Google Maps Marker documentation (kudos to
@hughvidos), map markers track view changes by default. This has meant
however that custom marker images or views on iOS and android, when
using Google Maps, cause high continuous CPU load.

The recommended use of tracksViewChanges is to enable when the marker
is being rendered or animated, then to disable it.

An example of use of tracksViewChanges would be to set it to false once
onMapReady has fired. This can be accomplished by toggling a state
property (e.g. `initialized`), and passing it to the child map markers
in their `tracksViewChanges` prop. Markers would then render correctly
on map load; CPU use falls off once the tracking ceases.