Adding onIdleCamera event for Google Maps
Calculation to get the bounds of the map northEast and southWest corners are not exact with the help of region data provided with deltas. Hence a new event is being added when the camera is idle at a position.

Usage - Wants to show the favourite places inside a particular region when user moves camera.