build: drop support for react-native < 0.63.4
We state that we support all versions, yet we "only" support > 0.51.
We shouldn't spend time supporting legacy versions of react-native.
If consumers are running versions < 0.63.4 they shouldn't expect to be
able to upgrade.
Furthermore, it should be possible to write functional components
such as Geojson which isn't possible with the current settings.

BREAKING CHANGE: drops supports for react-native < 0.63.4