fix(Android): animateToCamera to work before a map is laid out
Updating camera config doesn't work before a map is laid out. In that case,
keep the camera config to `cameraToSet` and apply it later.

To remove duplicate codes, moveToCamera is updated in this commit to accept
current CameraPosition to merge and duration for animation. animateToCamera
is updated to use moveToCamera.

Fixes #5162